Add Emergency Contact classes
Add annotations in EmergencyContact class which represent the
datamodel in the database. This class will be used to represent
Emergency Contacts in the program.

Add DAO for EmergencyContact. This will allow the application to
retrieve Emergency Contacts from the database.

Add GetEmergencyContact Servlet This Servlet will allow Emergency
Contacts to be retrieved from the database through a GET request.