Modify getByField() logic in Dao
Change logic of getByField() to return a List of all found records
matching the search string. This improves the method's functionality
as it enables the return of multiple records if more than one match
is found in the database.

Change calls to getByField() that expect only one return value to get
the first element of the returned list in order to preserve current
functionality.

Add getters and setters for EmergencyContact class to retrieve and
set values as necessary.

Modify comments to make them more clear and understandable.

Add EmergencyContact class to hibernate config file.