Fix register modal view logic
Move loading and view script from individual pages to navbar.jsp. This
ensures that an administrator can access the create account modal from
anywhere on the website.

Add permission checking to registermodal.jsp. This ensures that a user
has to belong to the administrator group to access the register modal.
This ensures that unauthorized users cannot create new accounts.