Gomezem/admin register only (#31)
* Removed sign up from login and added to nav

* Added the Create Account button to admin view only

* Add footer to login and fix comment

* Fix register modal view logic

Move loading and view script from individual pages to navbar.jsp. This
ensures that an administrator can access the create account modal from
anywhere on the website.

Add permission checking to registermodal.jsp. This ensures that a user
has to belong to the administrator group to access the register modal.
This ensures that unauthorized users cannot create new accounts.

* Reworded indications of sign up to create account

* Register forwards to prior page and cleaned up code

Co-authored-by: Jake Hansen <jacobdhansen@unomaha.edu>