Add emergency contact backend functionality (#30)
* Add Emergency Contact classes

Add annotations in EmergencyContact class which represent the
datamodel in the database. This class will be used to represent
Emergency Contacts in the program.

Add DAO for EmergencyContact. This will allow the application to
retrieve Emergency Contacts from the database.

Add GetEmergencyContact Servlet This Servlet will allow Emergency
Contacts to be retrieved from the database through a GET request.

* Modify getByField() logic in Dao

Change logic of getByField() to return a List of all found records
matching the search string. This improves the method's functionality
as it enables the return of multiple records if more than one match
is found in the database.

Change calls to getByField() that expect only one return value to get
the first element of the returned list in order to preserve current
functionality.

Add getters and setters for EmergencyContact class to retrieve and
set values as necessary.

Modify comments to make them more clear and understandable.

Add EmergencyContact class to hibernate config file.

* Complete logic for GetEmergencyContacts

Add functionality for GetEmergencyContacts to set request attribute
to matching Emergency Contacts from the database. This enables
the web application to retrieve all Emergency Contacts for an
authenticated user by requesting /getemergencycontacts.

Update comments for readability.

* Fix spelling error fieldName