Simplify GetAllLostAndInjured.java logic
Add status code names to UserStatus.java which serves to represent
the conversion from numerical status codes to text status codes.

Add helper method updateStatusNames() which updates the text status
code based on the numerical status code.

Add UserStatus object variable to User class in datamodel. This allows
a UserStatus object to be attached to a User object which greatly
simplifies logic throughout the webapp.

Update logic in firstresponders.jsp to display user status information
using the new backend logic to make JSP more readable.

Set toast positioning to fixed, so that when a user scrolls, the toast
notification is fixed within the view instead of scrolling with the
page.